# dynamically build the output schema