This addresses issue #230 by removing the `ArrowSchema` call in the physical plan execute function. It instead builds the output schema as it converts row groups. 

Benchmark with simulated network latency to storage
```
name                      old time/op    new time/op    delta
PersistentQuery/Types-10     638ms ± 2%     317ms ± 2%  -50.32%  (p=0.000 n=29+28)
PersistentQuery/Merge-10     2.51s ± 3%     2.22s ± 3%  -11.61%  (p=0.000 n=29+28)
PersistentQuery/Range-10     877ms ± 2%     562ms ± 3%  -35.91%  (p=0.000 n=30+28)

name                      old alloc/op   new alloc/op   delta
PersistentQuery/Types-10    68.3MB ± 0%    36.7MB ± 0%  -46.36%  (p=0.000 n=29+30)
PersistentQuery/Merge-10    32.0GB ± 3%    32.0GB ± 1%     ~     (p=0.296 n=30+28)
PersistentQuery/Range-10    2.43GB ± 1%    2.40GB ± 2%   -1.50%  (p=0.000 n=30+30)

name                      old allocs/op  new allocs/op  delta
PersistentQuery/Types-10     1.09M ± 0%     0.57M ± 0%  -47.27%  (p=0.000 n=30+30)
PersistentQuery/Merge-10     12.9M ± 1%     12.4M ± 1%   -4.16%  (p=0.000 n=29+30)
PersistentQuery/Range-10     2.86M ± 1%     2.34M ± 1%  -18.24%  (p=0.000 n=28+27)
```

Benchmark without the simulated network latency
```
name                      old time/op    new time/op    delta
PersistentQuery/Types-10    35.7ms ± 1%    17.1ms ± 1%  -51.95%  (p=0.000 n=22+28)
PersistentQuery/Merge-10     1.94s ± 7%     1.92s ± 5%     ~     (p=0.099 n=28+29)
PersistentQuery/Range-10     283ms ± 4%     268ms ± 5%   -5.41%  (p=0.000 n=29+29)

name                      old alloc/op   new alloc/op   delta
PersistentQuery/Types-10    68.2MB ± 0%    36.6MB ± 0%  -46.40%  (p=0.000 n=30+30)
PersistentQuery/Merge-10    32.0GB ± 2%    32.0GB ± 2%     ~     (p=0.781 n=29+30)
PersistentQuery/Range-10    2.42GB ± 1%    2.39GB ± 1%   -1.35%  (p=0.000 n=30+28)

name                      old allocs/op  new allocs/op  delta
PersistentQuery/Types-10     1.08M ± 0%     0.57M ± 0%  -47.30%  (p=0.000 n=29+30)
PersistentQuery/Merge-10     12.9M ± 0%     12.4M ± 1%   -4.01%  (p=0.000 n=30+30)
PersistentQuery/Range-10     2.85M ± 1%     2.33M ± 2%  -18.32%  (p=0.000 n=29+30)
```